cashback_bot/requirements.txt
aiosqlite>=0.18
python-telegram-bot>=20.0,<21.0
pillow>=10.0
pytesseract>=0.3.10
opencv-python-headless>=4.8
numpy>=1.24
pydantic>=1.10
pytz>=2023.3
